refactor: update struct names and visibility for consistency
