+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mix"
-version = "1.0.0"
+version = "1.0.1"
 edition = "2021"
 
 [dependencies]
