+++ src/commands/copy_snippet.rs
@@ -11,7 +11,7 @@ pub(crate) struct CopySnippet<'a> {
 }
 
 #[derive(Debug, Clone)]
-pub(crate) struct CopyOutput {
+pub struct CopyOutcome {
     pub key: String,
     pub relative_path: String,
     pub absolute_path: std::path::PathBuf,
@@ -22,14 +22,14 @@ impl CopySnippet<'_> {
         &self,
         storage: &SnippetStorage,
         clipboard: &dyn Clipboard,
-    ) -> Result<CopyOutput, AppError> {
+    ) -> Result<CopyOutcome, AppError> {
         let snippet = storage.resolve_snippet(self.query)?;
         let content = fs::read_to_string(&snippet.absolute_path)?;
         let project_root = find_project_root().ok();
         let expanded = expand_placeholders(&content, project_root.as_deref());
         clipboard.copy(expanded.as_ref())?;
 
-        Ok(CopyOutput {
+        Ok(CopyOutcome {
             key: snippet.key,
             relative_path: snippet.relative_path,
             absolute_path: snippet.absolute_path,
